pylint clevar/match_metrics/scaling/array_funcs.py
